Add more generalized save event
