+++ internal/v1_20_R4/pom.xml
@@ -54,6 +54,11 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
+      <artifactId>openinvcommon</artifactId>
+      <groupId>com.lishid</groupId>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
       <artifactId>openinvplugincore</artifactId>
       <groupId>com.lishid</groupId>
     </dependency>
