+++ internal/v1_20_R4/src/main/java/com/lishid/openinv/internal/v1_20_R4/OpenPlayer.java
@@ -16,6 +16,7 @@
 
 package com.lishid.openinv.internal.v1_20_R4;
 
+import com.lishid.openinv.event.OpenEvents;
 import com.mojang.logging.LogUtils;
 import net.minecraft.Util;
 import net.minecraft.nbt.CompoundTag;
@@ -91,6 +92,10 @@ public class OpenPlayer extends CraftPlayer {
 
     @Override
     public void saveData() {
+        if (OpenEvents.saveCancelled(this)) {
+            return;
+        }
+
         ServerPlayer player = this.getHandle();
         // See net.minecraft.world.level.storage.PlayerDataStorage#save(EntityHuman)
         try {
