+++ internal/v1_21_R1/pom.xml
@@ -54,6 +54,11 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
+      <artifactId>openinvcommon</artifactId>
+      <groupId>com.lishid</groupId>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
       <artifactId>openinvplugincore</artifactId>
       <groupId>com.lishid</groupId>
     </dependency>
